Them giao dien UI message va them cac Police va Elasticsearch

+++ Server/routes/api/v1.php
@@ -20,6 +20,7 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return 'Hello World';
 });
+
 Route::get('/search', [ProductController::class, 'search']);
 Route::get('/products/web', [ProductController::class, 'indexWeb']);
 Route::get('/products/web/{id}/', [ProductController::class, 'productDetail']);
@@ -48,7 +49,7 @@ Route::middleware(['auth:sanctum'])->group(function () {
     Route::get('/profile', [AuthController::class, 'profile']);
 
     Route::get('/orders/users', [OrderController::class, 'orderUser']);
-
+    Route::get('products/trash', [ProductController::class, 'productDelete'])->middleware('role:admin');
     Route::post('/momo/payment', [MoMoController::class, 'createPayment']);
 });
 
@@ -73,6 +74,8 @@ Route::middleware(['auth:sanctum', 'role:admin|manager'])->group(function () {
     Route::delete('/products/{id}', [ProductController::class, 'destroy']);
     Route::get('/products/{id}', [ProductController::class, 'show']);
     Route::put('/products/{id}', [ProductController::class, 'update']);
+  
+    Route::patch('/products/{id}/restore', [ProductController::class, 'restore']);
 
     Route::apiResource('attributes', AttributeController::class);
     Route::apiResource('attribute-options', AttributeOptionController::class);
@@ -98,28 +101,28 @@ Route::middleware(['auth:sanctum', 'role:admin|manager'])->group(function () {
     Route::get('/permissions', [RolePermissionController::class, 'indexPermissions']);
 
     // Tạo role & permission
-    Route::post('/roles', [RolePermissionController::class, 'createRole']);
-    Route::post('/permissions', [RolePermissionController::class, 'createPermission']);
+    Route::post('/roles', [RolePermissionController::class, 'createRole'])->middleware('role:admin');
+    Route::post('/permissions', [RolePermissionController::class, 'createPermission'])->middleware('role:admin');
 
     // Gán & xóa permission cho role
-    Route::post('/roles/assign-permissions', [RolePermissionController::class, 'assignPermissionToRole']);
-    Route::post('/roles/remove-permission', [RolePermissionController::class, 'removePermissionFromRole']);
+    Route::post('/roles/assign-permissions', [RolePermissionController::class, 'assignPermissionToRole'])->middleware('role:admin');
+    Route::post('/roles/remove-permission', [RolePermissionController::class, 'removePermissionFromRole'])->middleware('role:admin');
 
     // Gán & xóa role cho user
-    Route::post('/users/assign-role', [RolePermissionController::class, 'assignRoleToUser']);
-    Route::post('/users/remove-role', [RolePermissionController::class, 'removeRoleFromUser']);
+    Route::post('/users/assign-role', [RolePermissionController::class, 'assignRoleToUser'])->middleware('role:admin');
+    Route::post('/users/remove-role', [RolePermissionController::class, 'removeRoleFromUser'])->middleware('role:admin');
 
     // Xóa role & permission
-    Route::delete('/roles/{role}', [RolePermissionController::class, 'deleteRole']);
-    Route::delete('/permissions/{id}', [RolePermissionController::class, 'deletePermission']);
+    Route::delete('/roles/{role}', [RolePermissionController::class, 'deleteRole'])->middleware('role:admin');
+    Route::delete('/permissions/{id}', [RolePermissionController::class, 'deletePermission'])->middleware('role:admin');
 
     // Route::post('/roles/{id}/assign-all-permissions', [RolePermissionController::class, 'assignAllPermissionsToRole']);
 
-    Route::post('/roles/{id}/update-permissions', [RolePermissionController::class, 'updatePermissions']);
+    Route::post('/roles/{id}/update-permissions', [RolePermissionController::class, 'updatePermissions'])->middleware('role:admin');
 
-    Route::post('/roles/remove-all-permissions', [RolePermissionController::class, 'removeAllPermissionsFromRole']);
+    Route::post('/roles/remove-all-permissions', [RolePermissionController::class, 'removeAllPermissionsFromRole'])->middleware('role:admin');
 
-    Route::get('/roles/{id}/permissions', [RolePermissionController::class, 'getRolePermissions']);
+    Route::get('/roles/{id}/permissions', [RolePermissionController::class, 'getRolePermissions'])->middleware('role:admin');
 
     Route::patch('/products/{id}/restore', [ProductController::class, 'restore']);
 });
